consumer/requirements.txt
dependency-injector==4.41.0
kafka-python==2.0.2
mysqlclient==2.2.0
pytest==7.3.1
pytest-env==0.6.2
PyYAML==6.0
requests
SQLAlchemy==2.0.22